requirements.txt
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
fonttools==4.62.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.6
